feat(frontend): move target name update functionality to monitor settings
