+++ frontend/src/components/dashboard/monitor-settings.tsx
@@ -1,11 +1,10 @@
-import { AlertCircle, CheckCircle, Plus, Save, Settings } from "lucide-react"
+import { Plus, Settings } from "lucide-react"
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query"
-import { useEffect, useState } from "react"
+import { useState } from "react"
 import type { OfferMonitor } from "@/types/dashboard"
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card"
 import { Switch } from "@/components/ui/switch"
 import { Button } from "@/components/ui/button"
-import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
 import {
   Select,
@@ -15,7 +14,7 @@ import {
   SelectValue
 } from "@/components/ui/select"
 import { ConfigFormModal } from "@/components/notifications/config-form-modal"
-import { generateTelegramToken, listNotificationConfigs, toggleTargetNotifications, updateTargetName, updateTargetNotificationConfig } from "@/lib/api/sdk.gen"
+import { generateTelegramToken, listNotificationConfigs, toggleTargetNotifications, updateTargetNotificationConfig } from "@/lib/api/sdk.gen"
 
 interface MonitorSettingsProps {
   offerMonitor: OfferMonitor
@@ -25,16 +24,11 @@ interface MonitorSettingsProps {
 export default function MonitorSettings({ offerMonitor, isVisible }: MonitorSettingsProps) {
   const queryClient = useQueryClient()
   const [telegramBotUrl, setTelegramBotUrl] = useState<string | null>(null)
-  const [targetName, setTargetName] = useState(offerMonitor.name)
-  const [updateSuccess, setUpdateSuccess] = useState(false)
-  const [updateError, setUpdateError] = useState<string | null>(null)
   const [isModalOpen, setIsModalOpen] = useState(false)
 
   const selectedNotificationConfig = offerMonitor.notificationConfigId
 
-  useEffect(() => {
-    setTargetName(offerMonitor.name)
-  }, [offerMonitor.name])
+  
 
   const toggleNotificationsMutation = useMutation({
     mutationFn: (enable: boolean) =>
@@ -69,24 +63,7 @@ export default function MonitorSettings({ offerMonitor, isVisible }: MonitorSett
     },
   })
 
-  const updateNameMutation = useMutation({
-    mutationFn: (newName: string) => {
-      setUpdateError(null);
-      return updateTargetName({
-        path: { target_id: offerMonitor.id },
-        body: { name: newName },
-        throwOnError: true,
-      });
-    },
-    onSuccess: () => {
-      setUpdateSuccess(true);
-      queryClient.invalidateQueries({ queryKey: ['myTarget'] });
-      setTimeout(() => setUpdateSuccess(false), 2000);
-    },
-    onError: (err: any) => {
-      setUpdateError(err.message || "An unexpected error occurred.");
-    },
-  });
+  
 
   const { data: notificationConfigs } = useQuery({
     queryKey: ['notificationConfigs'],
@@ -123,50 +100,6 @@ export default function MonitorSettings({ offerMonitor, isVisible }: MonitorSett
         <CardDescription>Configure your offer monitoring preferences</CardDescription>
       </CardHeader>
       <CardContent className="space-y-6">
-        <div className="p-4 bg-gradient-to-r from-blue-50 to-indigo-50 rounded-lg">
-          <div className="flex items-center justify-between">
-            <div>
-              <h3 className="font-medium text-gray-900">Target Name</h3>
-              <p className="text-sm text-gray-600">Update the name of your scraping target.</p>
-            </div>
-            <div className="flex items-center space-x-2">
-              <Input
-                value={targetName}
-                onChange={(e) => {
-                  setTargetName(e.target.value)
-                  setUpdateError(null)
-                }}
-                className="w-48"
-                disabled={updateNameMutation.isPending}
-              />
-              <Button
-                onClick={() => updateNameMutation.mutate(targetName)}
-                disabled={updateNameMutation.isPending || (targetName === offerMonitor.name && !updateError) || updateSuccess}
-                className="w-32 justify-center bg-gradient-to-r from-violet-600 to-purple-600 hover:from-violet-700 hover:to-purple-700 transition-all duration-300 hover:scale-105"
-              >
-                {updateNameMutation.isPending ? (
-                  <>
-                    <div className="w-4 h-4 border-2 border-white/30 border-t-white rounded-full animate-spin mr-2" />
-                    Saving...
-                  </>
-                ) : updateSuccess ? (
-                  <CheckCircle className="w-5 h-5 text-white" />
-                ) : (
-                  <>
-                    <Save className="w-4 h-4 mr-2" />
-                    Update
-                  </>
-                )}
-              </Button>
-            </div>
-          </div>
-          {updateError && (
-            <div className="flex items-center mt-2 text-sm text-red-600">
-              <AlertCircle className="w-4 h-4 mr-2" />
-              {updateError}
-            </div>
-          )}
-        </div>
         <div className="flex items-center justify-between p-4 bg-gradient-to-r from-violet-50 to-purple-50 rounded-lg">
           <div>
             <h3 className="font-medium text-gray-900">Enable Notifications</h3>
